Add new forced photometry algorithms.
Each forced photometry algorithm has been moved to it's own file to
avoid confusion as they perform simliar, but different functions
to get the references/images and perform measurements.  I tried
to update the README with some documentation as well.
