+++ bin.src/forcedPhotCcdTemplateDia.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from lsst.dia.pipe.forcedPhotCcdTemplateDia import ForcedPhotCcdTemplateDiaTask
+ForcedPhotCcdTemplateDiaTask.parseAndRun()
