+++ bin.src/forcedPhotCoaddDiaSrc.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from lsst.dia.pipe.forcedPhotCoaddDiaSrc import ForcedPhotCoaddDiaSrcTask
+ForcedPhotCoaddDiaSrcTask.parseAndRun()
